src/test/java/com/upeu/connector/handler/BaseHandlerTest.java
package com.upeu.connector.handler;

import com.upeu.connector.DSpaceClient;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

/**
 * Test class for BaseHandler.
 */
public class BaseHandlerTest {

    private BaseHandler handler;
    private DSpaceClient mockClient;

    @BeforeEach
    public void setUp() {
        mockClient = mock(DSpaceClient.class); // Mock the DSpaceClient

        // Create a concrete implementation of BaseHandler for testing
        handler = new BaseHandler(mockClient) {
            @Override
            protected boolean validate(Object entity) {
                // Return true as a stub for validation
                return true;
            }
        };
    }

    @Test
    public void testGetClient() {
        // Verify that the client is correctly set
        assertNotNull(handler.getClient(), "Client should not be null");
        assertEquals(mockClient, handler.getClient(), "Client should match the mocked instance");
    }

    @Test
    public void testValidateEntity() {
        // Test the validate method
        Object entity = new Object(); // Example entity
        assertTrue(handler.validate(entity), "Validation should return true for this test case");
    }

    @Test
    public void testLogError() {
        // Test the logError method
        Exception e = new Exception("Test exception");
        assertDoesNotThrow(() -> handler.logError("Test message", e), "logError should not throw exceptions");
    }
}
